mw/requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-extensions==2.1.5
djangorestframework==3.9.1
djangorestframework-simplejwt==3.3
get==2018.11.19
idna==2.8
mysqlclient==1.4.1
post==2018.11.20
public==2018.11.20
PyJWT==1.7.1
pytz==2018.9
query-string==2018.11.20
request==2018.11.20
requests==2.21.0
six==1.12.0
soupsieve==1.8
urllib3==1.24.1
Werkzeug==0.14.1
